Fix modelling CI coverage and bundle budget
